服务器上一些修改，主要 stiger 完成 pubshm 数据统计 pop3 收信重置信件状态检查 一些界面修改 特殊模式 q 退出后光标尽量停留在普通模式同一篇文章 usermemo 扩容
git-svn-id: http://svn.kcn.cn/repos/kbs/trunk@9108 de4889a5-8ad4-4bf3-bd87-57997c21d515

+++ kbs_bbs/src/func.h
@@ -215,7 +215,7 @@ typedef ssize_t (*read_func)(int fd, void *buf, size_t count);
     unsigned int get_mailbox_prop(char *userid);
     unsigned int update_mailbox_prop(char *userid, unsigned int prop);
     int gen_title(const char *boardname );
-    off_t read_user_memo( char *userid, struct usermemo ** ppum );
+    int read_user_memo( char *userid, struct usermemo ** ppum );
 #if HAVE_MYSQL_SMTH == 1
 #ifdef BMSLOG
 	int bms_add(char *userid, char *boardname, time_t in, int out, char *memo );
